requirements.txt
youtube-transcript-api
langchain
langchain-openai
openai
langchain-community
python-dotenv
gradio
chromadb
tiktoken
pytube
